requirements.txt
numpy==2.1.2
pillow==11.0.0
pyfqmr==0.2.1
pygame==2.6.1
PyMetis==2023.1.1
PyOpenGL==3.1.7
scipy==1.14.1
